requirements.txt
attrs==23.1.0
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.4.post1
Flask==2.3.2
geopandas==0.11.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
platformdirs==3.10.0
polyline==2.0.0
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
Rtree==1.0.1
Shapely==1.8.5.post1
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
